+++ compass_app/app/lib/data/models/destination.dart
