+++ compass_app/app/lib/ui/results/presentation/results_screen.dart
@@ -1,6 +1,6 @@
-import 'package:compass_app/common/themes/colors.dart';
-import 'package:compass_app/features/results/presentation/result_card.dart';
-import 'package:compass_app/features/results/presentation/results_viewmodel.dart';
+import 'package:compass_app/ui/core/themes/colors.dart';
+import 'package:compass_app/ui/results/presentation/result_card.dart';
+import 'package:compass_app/ui/results/presentation/results_viewmodel.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
