+++ compass_app/app/lib/utils/result.dart
