+++ compass_app/app/test/data/repositories/destination/destination_repository_local_test.dart
@@ -1,6 +1,6 @@
 
-import 'package:compass_app/common/utils/result.dart';
-import 'package:compass_app/features/results/data/destination_repository_local.dart';
+import 'package:compass_app/utils/result.dart';
+import 'package:compass_app/data/repositories/destination/destination_repository_local.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
