+++ compass_app/app/test/ui/results/business/search_destination_usecase_test.dart
@@ -1,7 +1,7 @@
-import 'package:compass_app/common/utils/result.dart';
-import 'package:compass_app/features/results/business/model/destination.dart';
-import 'package:compass_app/features/results/business/usecases/search_destination_usecase.dart';
-import 'package:compass_app/features/results/data/destination_repository.dart';
+import 'package:compass_app/ui/results/business/search_destination_usecase.dart';
+import 'package:compass_app/utils/result.dart';
+import 'package:compass_app/data/models/destination.dart';
+import 'package:compass_app/data/repositories/destination/destination_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
